backend/requirements.txt
flask>=2.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
flasgger>=0.9.7
werkzeug>=2.0.0
